+++ src/components/common/Header/Header.module.scss
@@ -63,120 +63,3 @@
   height: 3.5rem;
   border-left: 1px solid $gray_D9;
 }
-
-.memberprofileGroup {
-  display: flex;
-  gap: -10rem;
-}
-
-.memberProfile {
-  @include text-m(600);
-
-  position: relative;
-  display: flex;
-  width: 3.8rem;
-  height: 3.8rem;
-  align-items: center;
-  justify-content: center;
-  border: 0.2rem solid $white_FF;
-  border-radius: 50%;
-  background-color: $gray_D9;
-  color: $white_FF;
-}
-
-.plusProfile {
-  @include text-m(600);
-
-  display: flex;
-  width: 3.8rem;
-  height: 3.8rem;
-  align-items: center;
-  justify-content: center;
-  border: 0.2rem solid $white_FF;
-  border-radius: 99px;
-  background-color: $blue_76;
-  color: $white_FF;
-}
-
-.myprofileGroup {
-  display: flex;
-  align-items: center;
-  padding: 0.4rem 1rem;
-  border-radius: 8px;
-
-  &:hover {
-    background-color: $gray_EE;
-  }
-}
-
-.myProfile {
-  @include text-m(600);
-
-  display: flex;
-  width: 3.8rem;
-  height: 3.8rem;
-  align-items: center;
-  justify-content: center;
-  border: 0.2rem solid $white_FF;
-  border-radius: 99px;
-  margin-right: 1rem;
-  background-color: $blue_76;
-  color: $white_FF;
-}
-
-.myName {
-  @include text-m(500);
-
-  display: flex;
-  align-items: center;
-  margin-right: 0.6rem;
-  color: $black_33;
-}
-
-.dropdown {
-  @include text-m(400);
-
-  position: absolute;
-  top: 6.4rem;
-  display: inline-flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 0.6rem;
-  border: 0.1rem solid $gray_D9;
-  border-radius: 0.8rem;
-  background: $white_FF;
-}
-
-.memberList {
-  margin: 0 0.6rem;
-  gap: 0.2rem;
-}
-
-.member {
-  display: flex;
-  align-items: center;
-  gap: 0.4rem;
-}
-
-.memberName {
-  @include text-m(400);
-}
-
-.nav {
-  display: flex;
-  flex-direction: column;
-  gap: 0.2rem;
-}
-
-.myNav {
-  display: flex;
-  width: 10rem;
-  height: 3rem;
-  align-items: center;
-  padding: 0 0.6rem;
-  border-radius: 8px;
-}
-
-.myNav:hover {
-  background-color: $gray_EE;
-}
