+++ src/components/common/Header/HeaderMemberProfile.module.scss
@@ -1,0 +1,62 @@
+.memberprofileGroup {
+  display: flex;
+  gap: -10rem;
+
+  .memberProfile {
+    @include text-m(600);
+
+    position: relative;
+    display: flex;
+    width: 3.8rem;
+    height: 3.8rem;
+    align-items: center;
+    justify-content: center;
+    border: 0.2rem solid $white_FF;
+    border-radius: 50%;
+    background-color: $gray_D9;
+    color: $white_FF;
+  }
+
+  .plusProfile {
+    @include text-m(600);
+
+    display: flex;
+    width: 3.8rem;
+    height: 3.8rem;
+    align-items: center;
+    justify-content: center;
+    border: 0.2rem solid $white_FF;
+    border-radius: 99px;
+    background-color: $blue_76;
+    color: $white_FF;
+  }
+}
+
+.dropdown {
+  @include text-m(400);
+
+  position: absolute;
+  top: 6.4rem;
+  display: inline-flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 0.6rem;
+  border: 0.1rem solid $gray_D9;
+  border-radius: 0.8rem;
+  background: $white_FF;
+}
+
+.memberList {
+  margin: 0 0.6rem;
+  gap: 0.2rem;
+
+  .member {
+    display: flex;
+    align-items: center;
+    gap: 0.4rem;
+  }
+
+  .memberName {
+    @include text-m(400);
+  }
+}
